Refactor: Organized import statements and fixed the next js 15+ params error
