+++ app/components/IssueStatus.tsx
@@ -1,6 +1,5 @@
 import { Status } from "@prisma/client";
 import { Badge } from "@radix-ui/themes";
-import React from "react";
 const statusMap: Record<
   Status,
   { label: string; color: "red" | "violet" | "green" }
