+++ app/components/NewIssueButton.tsx
@@ -1,6 +1,5 @@
 import { Button } from "@radix-ui/themes";
 import Link from "next/link";
-import React from "react";
 
 const NewIssueButton = () => {
   return (
