+++ app/components/index.ts
@@ -1,0 +1,6 @@
+export { default as Spinner } from "./Spinner";
+export { default as NewIssueButton } from "./NewIssueButton";
+export { default as Link } from "./Link";
+export { default as IssueStatus } from "./IssueStatus";
+export { default as IssueHeading } from "./IssueHeading";
+export { default as ErrorMessage } from "./ErrorMessage";
