+++ app/issues/[id]/page.tsx
@@ -1,17 +1,17 @@
-import IssueStatus from "@/app/components/IssueStatus";
+import { IssueStatus } from "@/app/components";
 import { prisma } from "@/prisma/prismaClient";
 import { Card, Flex, Heading } from "@radix-ui/themes";
 import { notFound } from "next/navigation";
 import Markdown from "react-markdown";
 
 interface Props {
-  params: { id: string };
+  params: Promise<{ id: string }>;
 }
 
 const IssueDetailsPage = async ({ params }: Props) => {
   const issue = await prisma.issue.findUnique({
     where: {
-      id: parseInt(params?.id),
+      id: parseInt((await params).id, 10),
     },
   });
